dashboard/backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pyyaml>=6.0.2
aiofiles>=24.1.0
markdown>=3.7
python-multipart>=0.0.9
openai>=1.76.0
pymediainfo>=6.1.0
